Fix CommonEventBusSubscriber mismatch
Mistakenly put world generation event into the newly created MultiblockEventBusSubscriber, instead of moving the onBlockBreak function check used for multiblock structure deconstruction.
